vera_backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic>=2.7.4
pydantic-settings>=2.0.3
openai==1.55.3
python-dotenv
httpx
sqlalchemy
psycopg2-binary
sentry-sdk
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart
pgvector
redis==5.0.1
websockets==12.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.74
langgraph-checkpoint>=2.1.1
langgraph-checkpoint-postgres>=2.0.23

# Integration dependencies
slack-sdk==3.27.1
jira==3.8.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
msgraph-sdk
azure-identity
requests-oauthlib==2.0.0
python-multipart==0.0.9
